Does it matter how many sutures were placed?  Should we remove that part? – For billing purposes, we do not need to know how many sutures were placed.
I do think we should have lengths (are these length ranges correct?  and is it ok to have a range or do you need a specific number?)

This also includes superficial closure with removal of particulate matter right?  If you’re asking about the lengths being the same for simple then yes, the lengths are the same for simple and intermediate.

so for deep sutures, 0, should be an option.   – Can you elaborate on this?  I’m not sure what you mean.